+++ simulations/losers-empirical/simulation.py
@@ -1,8 +1,6 @@
 import os
 import sys
 
-from numpy.random.mtrand import multivariate_normal
-import pandas as pd
 import numpy as np
 import pandas as pd
 from scipy.stats import multivariate_normal