+++ simulations/losers-presentation/variables.py
@@ -5,4 +5,4 @@ from .simulation import CONVENTIONAL, CONDITIONAL, HYBRID, PROJECTION
 N_SIMULATIONS = 1000
 
 estimators = [CONVENTIONAL, CONDITIONAL, HYBRID, PROJECTION]
-variables = product(range(N_SIMULATIONS), estimators, ("movers.csv", "oa.csv"))
+variables = product(range(N_SIMULATIONS), estimators, ("../losers-empirical/movers.csv",))